Fix Double -> BigDecimal
git-svn-id: svn://svn.opentaps.org/opentaps_all/versions/1.4/trunk@14463 d3523486-f5fe-0310-a41a-a15a4e76f3c7

+++ hot-deploy/opentaps-tests/src/org/opentaps/tests/analytics/tests/TestObjectGenerator.java
@@ -467,7 +467,7 @@ public class TestObjectGenerator {
                     callCtxt.put("productPricePurposeId", "PURCHASE");
                     callCtxt.put("currencyUomId", "USD");
                     callCtxt.put("fromDate", fromDate);
-                    callCtxt.put("price", new Double(getRandomIndex(200)));
+                    callCtxt.put("price", new BigDecimal(getRandomIndex(200)));
                     callCtxt.put("productStoreGroupId", "_NA_");
                     dispatcher.runSync("createProductPrice", callCtxt);
